requirements_extract.txt
gdtoolkit<4
godot_parser@git+https://github.com/raylu/godot_parser
Pillow
